add submission scripts to fusion pipeline and make more stringent
