+++ scripts/fusion_pipeline/2_4_fusion.py
@@ -1,55 +1,27 @@
-"""Fusion Part 2.4: Threshold and skeltonize the images on each scale."""
+"""Fusion Part 2.4: Threshold and skeletonize the images on each scale."""
 
-##################################################################################################
-#                                           IMPORTS
-##################################################################################################
 import argparse
 import time
 
 import dask.array as da
+import numpy as np
+import zarr
 from skimage.morphology import skeletonize as sk_skeletonize
 
 from skeleplex.skeleton._skeletonize import threshold_skeleton
+from skeleplex.utils._chunked import iteratively_process_chunks_3d
 
-##################################################################################################
-#                                           RUNNING FUSION
-##################################################################################################
+from ._constants import IMAGE_PREFIX, THRESHOLDS
 
-# Define the image prefix used to name the files
-image_prefix = "LADAF-2021-17-left-v7_processed"  # ADAPT HERE
+image_prefix = IMAGE_PREFIX
 
+# Thresholds per scale for binary skeleton generation
+thresholds = THRESHOLDS
 
-# Example: define scales and their valid ranges
-scale_ranges_manual = {
-    -1: (1, 10),
-    -3: (10, 250),
-}  # ADAPT HERE
-
-# define threshold for the final fusion 3 step
-thresholds = {
-    -1: 0.55,
-    -3: 0.7,
-}
-
-# Load the initial image (here: label)
-lung_image = da.from_zarr(f"/data/{image_prefix}.zarr")  # ADAPT HERE
-
-
-################ Fusion Part 2.4 ################
 parser = argparse.ArgumentParser()
-parser.add_argument(
-    "--job-index", help="this is the index of the submitted job", type=int
-)
-parser.add_argument(
-    "--job-index-offset",
-    help="this number assists in getting the negative "
-    "and positive scales required for the fusion algorithm (submit a positive integer)",
-    type=int,
-)
-parser.add_argument("--workers", help="this sets the number of workers", type=int)
-
+parser.add_argument("--job-index", type=int)
+parser.add_argument("--job-index-offset", type=int)
 args = parser.parse_args()
-
 
 scale_number = args.job_index - args.job_index_offset
 print("Job Index: ", args.job_index)
@@ -58,36 +30,42 @@ print("Scale number: ", scale_number)
 threshold = thresholds[scale_number]
 print("Threshold: ", threshold)
 
-re_scale_factor = 1 / (2 ** (scale_number))
-
-# load skeleton prediction image
 skel_pred_image = da.from_zarr(
     f"/data/{image_prefix}_skeleton_predictions_on_scales.zarr/scale{scale_number}"
 )
-
-# load the scaled image to mask the skeleton prediction image
 scaled_image = da.from_zarr(
     f"/data/{image_prefix}_image_scaled.zarr/scale{scale_number}"
 )
 
-# mask out the background in the skeleton prediction image
-time_start_masking = time.time()
-mask = scaled_image > 0
-masked_segmentation = da.where(mask, skel_pred_image, 0)
 
-masked_segmentation.to_zarr(
+def _mask_fn(skel: np.ndarray, seg: np.ndarray) -> np.ndarray:
+    return np.where(seg > 0, skel, 0).astype(skel.dtype)
+
+
+# Mask out the background in the skeleton prediction image
+time_start_masking = time.time()
+save_masked = zarr.open(
     f"/data/{image_prefix}_skeleton_predictions_on_scales.zarr/scale{scale_number}_masked",
     mode="w",
+    shape=skel_pred_image.shape,
+    chunks=(192, 192, 192),
+    dtype=skel_pred_image.dtype,
 )
-# print(f"--- Masking the skeleton prediction image
-# with the scaled segmentation took {time.time() - time_start_masking} seconds ---")
+
+iteratively_process_chunks_3d(
+    input_arrays=(skel_pred_image, scaled_image),
+    output_zarr=save_masked,
+    function_to_apply=_mask_fn,
+    chunk_shape=(192, 192, 192),
+    extra_border=(0, 0, 0),
+)
+print(f"--- Masking took {time.time() - time_start_masking} seconds ---")
 
 masked_segmentation_reloaded = da.from_zarr(
     f"/data/{image_prefix}_skeleton_predictions_on_scales.zarr/scale{scale_number}_masked"
 )
 
-
-# Threshold Image
+# Threshold image
 start_time3 = time.time()
 lung_image_binary_skeleton = threshold_skeleton(
     masked_segmentation_reloaded, threshold=threshold
@@ -99,26 +77,27 @@ lung_image_binary_skeleton.to_zarr(
 lung_image_binary_skeleton_reloaded = da.from_zarr(
     f"/data/{image_prefix}_skeleton_predictions_on_scales.zarr/scale{scale_number}_ts"
 )
-print(f"--- Thresholding optimal tree took {time.time() - start_time3} seconds ---")
+print(f"--- Thresholding took {time.time() - start_time3} seconds ---")
 
-
-# Perform Thinning / Skeletonizing
+# Thinning / Skeletonizing
 start_time4 = time.time()
 lung_image_binary_skeleton_reloaded = lung_image_binary_skeleton_reloaded.rechunk(
     (192, 192, 192)
 )
-lung_image_skeletonized = lung_image_binary_skeleton_reloaded.map_overlap(
-    sk_skeletonize,
-    depth=10,
-    boundary=0,
+
+save_skel = zarr.open(
+    f"/data/{image_prefix}_skeletonized_on_scales.zarr/scale{scale_number}",
+    mode="w",
+    shape=lung_image_binary_skeleton_reloaded.shape,
+    chunks=(192, 192, 192),
     dtype=lung_image_binary_skeleton_reloaded.dtype,
 )
 
-lung_image_skeletonized = lung_image_skeletonized.rechunk((192, 192, 192))
-lung_image_skeletonized.to_zarr(
-    f"/data/{image_prefix}_skeletonized_on_scales.zarr/scale{scale_number}",
-    mode="w",
+iteratively_process_chunks_3d(
+    input_arrays=lung_image_binary_skeleton_reloaded,
+    output_zarr=save_skel,
+    function_to_apply=sk_skeletonize,
+    chunk_shape=(192, 192, 192),
+    extra_border=(10, 10, 10),
 )
-
-# print(f"--- Skeletonizing the thresholded
-# image took {time.time() - start_time4} seconds ---")
+print(f"--- Skeletonizing took {time.time() - start_time4} seconds ---")
