+++ scripts/fusion_pipeline/3_fusion.py
@@ -1,141 +1,34 @@
-"""Fusion Part 3.1: Generate Optimal Tree from Scale Mapped Skeleton Predictions."""
+"""Fusion Part 3: Generate Optimal Tree from Scale Mapped Skeleton Predictions."""
 
-##################################################################################################
-#                                           IMPORTS
-##################################################################################################
 import time
 
 import dask.array as da
-import numpy as np
+import zarr
+from scripts.fusion_pipeline._constants import IMAGE_PREFIX, SCALE_RANGES_MANUAL
 from skimage.morphology import skeletonize as sk_skeletonize
 
 from skeleplex.skeleton import repair_breaks_lazy
-
-##################################################################################################
-#                                           FUNCTIONS
-##################################################################################################
-
-
-def fused_tree_generator(
-    scale_map: np.ndarray,
-    image: np.ndarray,
-    scale_ranges: dict,
-    scaled_images_dict: dict,
-) -> np.ndarray:
-    """
-    Generate Optimal Tree.
-
-    Fuse the re-scaled skeleton predictions together to
-    form the "optimal tree". The fused skeleton prediction
-    ("optimal tree") is generated by using the scale map
-    to guide which scale's skeleton prediction to
-    use at each voxel.
-
-    Parameters
-    ----------
-    scale_map : np.ndarray
-        Array with non-zero values documenting the scale of
-        each voxel in the segmentation.
-    image : np.array
-        original image that serves as the mask guide
-    scale_ranges : dict
-        This dictionary is used to map the scales to the radii
-        in the radius_map.
-    scaled_images_dict : dict
-        Dictionary containing all re-scaled skeleton predictions.
-
-    Returns
-    -------
-    np.ndarray
-        Array of same shape as input image, containing
-        scale mapped values.
-
-    """
-    fused_tree = da.zeros_like(image, dtype=image.dtype)
-
-    for key in scale_ranges.keys():
-        scale_number = key
-        print("Processing scale number: ", scale_number)
-
-        mask = scale_map == scale_number
-        name = f"{image_prefix}_image_skeletonized_rescaled_from_{scale_number}"
-        print(name)
-
-        scaled_images_dict[name] = scaled_images_dict[name].rechunk(scale_map.chunks)
-        fused_tree = da.where(mask, scaled_images_dict[name], fused_tree)
-
-    fused_tree.to_zarr(f"/data/{image_prefix}_fused_tree.zarr", mode="w")
-    return fused_tree
-
-
-##################################################################################################
-#                                           RUNNING FUSION
-##################################################################################################
+from skeleplex.skeleton.fusion.scale_image import pad_to_match
+from skeleplex.skeleton.fusion.tree_fusion import fused_tree_generator
+from skeleplex.utils._chunked import iteratively_process_chunks_3d
 
 # Define the image prefix used to name the files
-image_prefix = "LADAF-2021-17-left-v7_processed"  # ADAPT HERE
-
+image_prefix = IMAGE_PREFIX
 # Example: define scales and their valid ranges
-scale_ranges_manual = {
-    -1: (1, 10),
-    -3: (10, 150),
-}  # ADAPT HERE
+scale_ranges_manual = SCALE_RANGES_MANUAL
 
-
-# Load the initial image (here: label)
 lung_image = da.from_zarr(f"/data/{image_prefix}.zarr")  # ADAPT HERE
 lung_image = lung_image.rechunk((192, 192, 192))
 print("Lung image shape: ", lung_image.shape)
 
-
-################ From fusion Part 1 ################
 start_time1 = time.time()
 
-# Load Scale Map
 lung_image_scale_map = da.from_zarr(
     f"/data/{image_prefix}_image_scale_map_processed.zarr/scale_original"
 )
 lung_image_scale_map = lung_image_scale_map.rechunk((192, 192, 192))
 print("Lung image scale map shape: ", lung_image_scale_map.shape)
 
-
-################ From fusion Part 2 ################
-
-
-def pad_to_match(Original_Image, rescaled_image, value=0):
-    """Pad rescaled image to match the shape of the original image.
-
-    Parameters
-    ----------
-    Original_Image : np.ndarray
-        The original image that serves as the reference for padding.
-    rescaled_image : np.ndarray
-        The rescaled image that needs to be padded to match
-        the shape of the original image.
-    value : int, optional
-        The value to use for padding, by default 0.
-    """
-    if Original_Image.ndim != rescaled_image.ndim:
-        raise ValueError(
-            "Original_Image and rescaled_image must have the same number of dimensions"
-        )
-
-    pad_width = []
-    for a, b in zip(Original_Image.shape, rescaled_image.shape, strict=False):
-        diff = a - b
-        if diff < 0:
-            raise ValueError(
-                "rescaled_image must not be larger than Original_Image in any dimension"
-            )
-
-        before = diff // 2
-        after = diff - before
-        pad_width.append((before, after))
-
-    return da.pad(rescaled_image, pad_width, mode="constant", constant_values=value)
-
-
-# Load rescaled images to combine rescaled images via scale map
 multiscale_images = {}
 for key in scale_ranges_manual.keys():
     scale_number = key
@@ -152,46 +45,46 @@ for key in scale_ranges_manual.keys():
 
 print(f"--- Loading all images took {time.time() - start_time1} seconds ---")
 
-################ Fusion Part 3 ################
+# Combine rescaled images via scale map to generate optimal tree
 start_time2 = time.time()
-# Combine resclaed images to one via scale map: generate optimal tree
-lung_image_optimum = fused_tree_generator(
+fused_tree_generator(
     lung_image_scale_map, lung_image, scale_ranges_manual, multiscale_images
 )
-lung_image_optimum = lung_image_optimum.rechunk((192, 192, 192))
 print(f"--- Generating optimal tree took {time.time() - start_time2} seconds ---")
 
-
-# Load Optimal Tree
 lung_image_optimum = da.from_zarr(f"/data/{image_prefix}_fused_tree.zarr")
-
 
 # Repair breaks in the final skeleton
 start_time5 = time.time()
-lung_image_repaired = repair_breaks_lazy(
+repair_breaks_lazy(
     skeleton_path=f"/data/{image_prefix}_fused_tree.zarr",
     segmentation_path=f"/data/{image_prefix}.zarr",
-    output_path=f"/data/{image_prefix}_final_skeletonzarr",
+    output_path=f"/data/{image_prefix}_final_skeleton.zarr",
     repair_radius=40,
     chunk_shape=(256, 256, 256),
     backend="cupy",
 )
 
-# print( f"--- Repairing breaks in final
-# skeleton took {time.time() - start_time5} seconds ---")
-
-
-# perform thinning / skeletonizing on the repaired skeleton
+# Thinning / Skeletonizing on the repaired skeleton
 start_time6 = time.time()
 lung_image_repaired = da.from_zarr(f"/data/{image_prefix}_final_skeleton.zarr")
 lung_image_repaired = lung_image_repaired.rechunk((192, 192, 192))
-lung_image_skeletonized_final = lung_image_repaired.map_overlap(
-    sk_skeletonize, depth=10, boundary=0, dtype=lung_image_repaired.dtype
+
+save_here = zarr.open(
+    f"/data/{image_prefix}_final_skeleton_skeletonized.zarr",
+    mode="w",
+    shape=lung_image_repaired.shape,
+    chunks=(192, 192, 192),
+    dtype=lung_image_repaired.dtype,
 )
 
-lung_image_skeletonized_final.to_zarr(
-    f"/data/{image_prefix}_final_skeleton_skeletonized.zarr", mode="w"
+iteratively_process_chunks_3d(
+    input_arrays=lung_image_repaired,
+    output_zarr=save_here,
+    function_to_apply=sk_skeletonize,
+    chunk_shape=(192, 192, 192),
+    extra_border=(10, 10, 10),
 )
-# print(f"--- Skeletonizing final skeleton
-#  took {time.time() - start_time6} seconds ---")
+print(f"--- Skeletonizing final skeleton took {time.time() - start_time6} seconds ---")
+
 print("Finished Fusion Part 3: Generated Fused Tree and final skeleton.")
