+++ src/skeleplex/skeleton/fusion/tree_fusion.py
@@ -1,0 +1,58 @@
+import dask.array as da
+import numpy as np
+
+
+def fused_tree_generator(
+    scale_map: np.ndarray,
+    image: np.ndarray,
+    scale_ranges: dict,
+    scaled_images_dict: dict,
+    image_prefix: str
+) -> np.ndarray:
+    """
+    Generate Optimal Tree.
+
+    Fuse the re-scaled skeleton predictions together to
+    form the "optimal tree". The fused skeleton prediction
+    ("optimal tree") is generated by using the scale map
+    to guide which scale's skeleton prediction to
+    use at each voxel.
+
+    Parameters
+    ----------
+    scale_map : np.ndarray
+        Array with non-zero values documenting the scale of
+        each voxel in the segmentation.
+    image : np.array
+        original image that serves as the mask guide
+    scale_ranges : dict
+        This dictionary is used to map the scales to the radii
+        in the radius_map.
+    scaled_images_dict : dict
+        Dictionary containing all re-scaled skeleton predictions.
+    image_prefix : str
+        Prefix for the image files.
+
+
+    Returns
+    -------
+    np.ndarray
+        Array of same shape as input image, containing
+        scale mapped values.
+
+    """
+    fused_tree = da.zeros_like(image, dtype=image.dtype)
+
+    for key in scale_ranges.keys():
+        scale_number = key
+        print("Processing scale number: ", scale_number)
+
+        mask = scale_map == scale_number
+        name = f"{image_prefix}_image_skeletonized_rescaled_from_{scale_number}"
+        print(name)
+
+        scaled_images_dict[name] = scaled_images_dict[name].rechunk(scale_map.chunks)
+        fused_tree = da.where(mask, scaled_images_dict[name], fused_tree)
+
+    fused_tree.to_zarr(f"/data/{image_prefix}_fused_tree.zarr", mode="w")
+    return fused_tree
